Some NC paint research
Also add guidelines to editor.config.

+++ src/thirtytwo/Win32/Graphics/Gdi/HRGN.cs
@@ -10,6 +10,11 @@ public unsafe partial struct HRGN : IDisposable
 {
     public void Dispose()
     {
+        if (IsFull)
+        {
+            return;
+        }
+
         if (!IsNull)
         {
             Interop.DeleteObject(this);
@@ -17,6 +22,18 @@ public unsafe partial struct HRGN : IDisposable
 
         Unsafe.AsRef(in this) = default;
     }
+
+    /// <summary>
+    ///  Special <see cref="HRGN"/> sent during WM_NCPAINT to indicate the entire window.
+    /// </summary>
+    public static HRGN Full { get; } = (HRGN)1;
+
+    /// <summary>
+    ///  Is special <see cref="HRGN.Full"/> value.
+    /// </summary>
+    public bool IsFull => Value == Full.Value;
+
+    // There is also a special HRGN_MONITOR (2) that is used when maximized. Not sure if this escapes.
 
     public static HRGN FromRectangle(Rectangle rectangle) =>
         Interop.CreateRectRgn(rectangle.X, rectangle.Y, rectangle.Right, rectangle.Bottom);
@@ -30,13 +47,27 @@ public unsafe partial struct HRGN : IDisposable
 
     public static HRGN CreateEmpty() => Interop.CreateRectRgn(0, 0, 0, 0);
 
-    public static HRGN CombineRegion(HRGN first, HRGN second, RegionCombineMode combineMode) =>
-        CombineRegion(first, second, combineMode, out _);
+    public HRGN Combine(HRGN region, RegionCombineMode combineMode) =>
+        Combine(region, combineMode, out _);
 
-    public static HRGN CombineRegion(HRGN first, HRGN second, RegionCombineMode combineMode, out RegionType type)
+    public HRGN Combine(HRGN region, RegionCombineMode combineMode, out RegionType type)
     {
         HRGN hrgn = CreateEmpty();
-        type = (RegionType)Interop.CombineRgn(hrgn, first, second, (RGN_COMBINE_MODE)combineMode);
+        type = (RegionType)Interop.CombineRgn(hrgn, this, region, (RGN_COMBINE_MODE)combineMode);
+        if (type == RegionType.Error)
+        {
+            hrgn.Dispose();
+        }
+
+        return hrgn;
+    }
+
+    public HRGN Copy() => Copy(out _);
+
+    public HRGN Copy(out RegionType type)
+    {
+        HRGN hrgn = CreateEmpty();
+        type = (RegionType)Interop.CombineRgn(hrgn, this, default, (RGN_COMBINE_MODE)RegionCombineMode.Copy);
         if (type == RegionType.Error)
         {
             hrgn.Dispose();
